fix before and after all
