+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/LeonardoTestUtils.scala
@@ -654,99 +654,6 @@ trait LeonardoTestUtils
       }
     }
   }
-  def withRawlsWorkspace[T](
-    projectName: AzureBillingProjectName
-  )(testCode: WorkspaceResponse => T)(implicit authToken: AuthToken): T = {
-    // hardcode this if you want to use a static workspace
-//    val workspaceName = "ddf3f5fa-a80e-4b2f-ab6e-9bd07817fad1-azure-test-workspace"
-
-    val workspaceName = generateWorkspaceName()
-
-    println(s"withRawlsWorkspace: Calling create rawls workspace with name ${workspaceName}")
-
-    Rawls.workspaces.create(
-      projectName.value,
-      workspaceName,
-      Set.empty,
-      Map("disableAutomaticAppCreation" -> "true")
-    )
-
-    val response =
-      workspaceResponse(projectName.value, workspaceName).unsafeRunSync()(cats.effect.unsafe.IORuntime.global)
-
-    println(s"withRawlsWorkspace: Rawls workspace create called, response: ${response}")
-
-    testCode(response)
-  }
-
-  private def workspaceResponse(projectName: String, workspaceName: String)(implicit
-    authToken: AuthToken
-  ): IO[WorkspaceResponse] = for {
-    responseString <- IO.pure(Rawls.workspaces.getWorkspaceDetails(projectName, workspaceName))
-    json <- IO.fromEither(parser.parse(responseString))
-    parsedResponse <- IO.fromEither(json.as[WorkspaceResponse])
-  } yield parsedResponse
-
-  case class WorkspaceResponse(accessLevel: Option[String],
-                               canShare: Option[Boolean],
-                               canCompute: Option[Boolean],
-                               catalog: Option[Boolean],
-                               workspace: WorkspaceDetails,
-                               owners: Option[Set[String]],
-                               azureContext: Option[AzureCloudContext]
-  )
-
-  implicit val azureContextDecoder: Decoder[AzureCloudContext] = Decoder.instance { c =>
-    for {
-      tenantId <- c.downField("tenantId").as[String]
-      subscriptionId <- c.downField("subscriptionId").as[String]
-      resourceGroupId <- c.downField("managedResourceGroupId").as[String]
-    } yield AzureCloudContext(TenantId(tenantId),
-                              SubscriptionId(subscriptionId),
-                              ManagedResourceGroupName(resourceGroupId)
-    )
-  }
-
-  implicit val workspaceDetailsDecoder: Decoder[WorkspaceDetails] = Decoder.forProduct10(
-    "namespace",
-    "name",
-    "workspaceId",
-    "bucketName",
-    "createdDate",
-    "lastModified",
-    "createdBy",
-    "isLocked",
-    "billingAccountErrorMessage",
-    "errorMessage"
-  )(WorkspaceDetails.apply)
-
-  implicit val workspaceResponseDecoder: Decoder[WorkspaceResponse] = Decoder.forProduct7(
-    "accessLevel",
-    "canShare",
-    "canCompute",
-    "catalog",
-    "workspace",
-    "owners",
-    "azureContext"
-  )(WorkspaceResponse.apply)
-
-  // Note this isn't the full model available, we only need a few fields
-  // The full model lives in rawls here https://github.com/broadinstitute/rawls/blob/develop/model/src/main/scala/org/broadinstitute/dsde/rawls/model/WorkspaceModel.scala#L712
-  case class WorkspaceDetails(
-    namespace: String,
-    name: String,
-    workspaceId: String,
-    bucketName: String,
-    createdDate: String,
-    lastModified: String,
-    createdBy: String,
-    isLocked: Boolean = false,
-    billingAccountErrorMessage: Option[String] = None,
-    errorMessage: Option[String] = None
-  )
-
-  private def generateWorkspaceName(): String =
-    s"${UUID.randomUUID().toString()}-azure-test-workspace"
 
   def generateAzureDiskName(): String =
     s"automation-test-disk-${UUID.randomUUID().toString().substring(0, 8)}"
